+++ internal/helmutil/repo_entry_test.go
@@ -111,7 +111,6 @@ func TestLookupRepoEntry(t *testing.T) {
 	}
 
 	for name, tc := range testCases {
-		tc := tc
 		t.Run(name, func(t *testing.T) {
 			teardown := tc.setup()
 			defer teardown()